+++ src/or/rendcommon.c
@@ -297,7 +297,7 @@ int rend_cache_store(const char *desc, size_t desc_len)
   e->desc = tor_malloc(desc_len);
   memcpy(e->desc, desc, desc_len);
 
-  log_fn(LOG_INFO,"Successfully stored rend desc '%s', len %d", query, desc_len);
+  log_fn(LOG_INFO,"Successfully stored rend desc '%s', len %d", query, (int)desc_len);
   return 0;
 }
